refactor: introduce Ollama client abstraction
